Add test for Huge MSBFS ASP as well

+++ algo/src/main/java/org/neo4j/graphalgo/impl/msbfs/HugeMultiSourceBFS.java
@@ -177,7 +177,7 @@ public final class HugeMultiSourceBFS implements Runnable, MsBFSAlgo {
             }
         }
 
-        long depth = 0;
+        int depth = 0;
 
         while (true) {
             long nodeId = -1L;
